+++ lib/src/elements/filter/encoder/video/vulkan/mod.rs
@@ -1,0 +1,3 @@
+mod vulkan_encoder;
+
+pub use vulkan_encoder::{VulkanCodec, VulkanEncoder, VulkanEncoderError, VulkanEncoderOptions};